Creating a new workspace to handle as the virtualized rom manager.
Rom manager's plans are to attach to the emulator and act as a place to
hot swappable roms, "Hopefully" feat: Add workspace and cartridge-rs
package

This commit introduces a new workspace to manage the virtualized ROM
manager. It also includes the new `cartridge-rs` package, which will
handle the display functionality.